+++ src/scene/Scene.cpp
@@ -599,10 +599,10 @@ EntityVisibility getEntityVisibility(Entity & entity, bool cullingOnly) {
 	
 	// If the head/center is not visible test all other groups
 	for(const VertexGroup & group : entity.obj->grouplist) {
-		if(ObjVertHandle(group.origin) != entity.obj->fastaccess.head_group_origin &&
-		   isPointVisible(entity.obj->vertexWorldPositions[group.origin].v, &entity)) {
+		if(ObjVertHandle(size_t(group.origin)) != entity.obj->fastaccess.head_group_origin &&
+		   isPointVisible(entity.obj->vertexWorldPositions[size_t(group.origin)].v, &entity)) {
 			if(entity.obj->fastaccess.head_group_origin == ObjVertHandle() &&
-			   isPointInViewCenter(entity.obj->vertexWorldPositions[group.origin].v)) {
+			   isPointInViewCenter(entity.obj->vertexWorldPositions[size_t(group.origin)].v)) {
 				return EntityInFocus;
 			}
 			return EntityVisible;
